pyproject.toml
[tool.poetry]
name = "Task2-api"
version = "0.1.0"
description = "Task2 API"
authors = ["Antoni Szych"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "~5.2.0"
djangorestframework = "^3.15.0"
psycopg = {extras = ["binary"], version = "^3.1.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^9.0"
pytest-django = "^4.12"
factory-boy = "^3.3"
black = "^26"
pytest-factoryboy = "^2.8.1"
isort = "^8.0.1"
mypy = "^1.19"
pytest-mock = "^3.15.1"
django-stubs = {extras = ["compatible-mypy"], version = "^5.2.9"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.16.8"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task2_api.settings"
python_files = "test_*.py"

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "task2_api.settings"
